Update : Added authentication to product out db

+++ app/Inventory_in/model.py
@@ -2,14 +2,14 @@ import DateTime
 import uuid
 from datetime import datetime, timezone
 from database.database import Base
-from sqlalchemy import JSON, Boolean, Column, Integer, Date, String,DateTime, Uuid, true
+from sqlalchemy import JSON, Boolean, Column, Integer, Date, String,DateTime, Uuid, Float
 from sqlalchemy.orm import Relationship
 
 class InventoryDB(Base):
     __tablename__ = "inventory"
     invoice_id = Column(String, primary_key=True, default=str(uuid.uuid4()))
     invoice_number = Column(String)
-    invoice_amount = Column(Integer)
+    invoice_amount = Column(Float)
     invoice_date = Column(Date)
     inventory_paydate = Column(Date)
     vendor = Column(String)
